Rename JUnit5 provider to JUnit Platform provider, as suggested by @marcphilipp
