+++ surefire-providers/surefire-junit-platform/src/main/resources/META-INF/services/org.apache.maven.surefire.providerapi.SurefireProvider
@@ -1,0 +1,1 @@
+org.apache.maven.surefire.junitplatform.JUnitPlatformProvider
