--- surefire-providers/surefire-junit5/src/main/resources/META-INF/services/org.apache.maven.surefire.providerapi.SurefireProvider
@@ -1,1 +1,0 @@
-org.apache.maven.surefire.junit5.JUnitPlatformProvider
